% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function cycpar.R
\name{cycpar}
\alias{cycpar}
\title{Computes the cyclic parameters from parameters linear model}
\usage{
cycpar(a1, a2, P)
}
\arguments{
\item{a1}{b1*cos((2pi/P)*b2) ,     parameter of term cos((2pi/P)*T)}

\item{a2}{b1*sin((2pi/P)*b2) ,     parameter of term sin((2pi/P)*T)}

\item{P}{is the periodicity of the cycle}
}
\value{
The amplitude and phase of the cyclic process
}
\description{
This function computes the cyclic parameters from parameters linear model (see Verboon & Leontjevas, 2018)
}
\examples{
cycpar(a1,a2,P)
}
\keyword{cyclic}
\keyword{models}
