% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function fitCyclic.R
\name{fitCyclic}
\alias{fitCyclic}
\title{Fits the cyclic model, using a simple linear model.}
\usage{
fitCyclic(dat, yvar, xvar, dayNumber = NULL, cov = NULL, P = NULL,
  ymin = -1, ymax = 1, step = 0.25)
}
\arguments{
\item{yvar}{dependent variable}

\item{xvar}{time variabele}

\item{dayNumber}{indicates the days, used in the plots. If NULL plot assumes only one day.}

\item{cov}{vector of names containing additional variabeles (e.g. cov = c("x1", "daynr"))}

\item{P}{the periodicity of the cycle. If NULL the maximum of xvar will be taken.}

\item{ymin, ymax, steps}{parameters that control axes of the plot}
}
\value{
list containing the following elements:

parameters =      amplitude and phase

fit =             object from lm()

rawDataPlot =     plot of predicted values for all observations

meansPlot =       plot with predictions averaged over subjects,

oneCyclePlot =    plot with predictions for one cycle,
}
\description{
This function fits the cyclic model, using a simple linear model (see Verboon & Leontjevas, 2018)
}
\examples{
fitCyclic(dat, yvar = y = "dependentVar", xvar = "beepnr", dayNumber = "daynr")
}
\keyword{ESM}
\keyword{cyclic}
\keyword{model}
